improved rpm calculation, status updates, rotary encoder
* Changed the RPM calculation function to utilize millis() to calculate
the amount of interrupts per second, instead of activating interupts
during a set delay. This gives the same accuracy, but doesn't freeze the
rest of the main loop.
* Made the status over serial more modular and added a delay function
to not spam status to often.
* Increased default PWM duty cycle to 50 %
* Added a rotary encoder to manually control the speed of the fan.
The readings are still a bit flaky/sensitive, but this is most likely
due to to frequent readings (polling, no interrupts) and the encoder
itself being imprecise while rotating. But it works for now.
* Pressing the encoder's button resets fan speed back to default.